ТЗ-9 первая итерация.
